requirements.txt
# Core (installed via pip install -e .)
torch>=2.0
torchvision>=0.15
open-clip-torch>=2.24
Pillow>=10.0
matplotlib>=3.7
scikit-learn>=1.3
tqdm>=4.65

# Scripts
bing-image-downloader>=1.1
requests>=2.31